+++ cmd/mdmctl/apply_dep_profile.go
@@ -2,20 +2,43 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"encoding/json"
 	"flag"
 	"fmt"
+	"net/url"
 	"os"
 
 	"github.com/micromdm/dep"
+	"github.com/micromdm/micromdm/crypto"
 	"github.com/pkg/errors"
 )
+
+func certificatesFromURL(serverURL string, insecure bool) ([]*x509.Certificate, error) {
+	urlParsed, err := url.Parse(serverURL)
+	if err != nil {
+		return nil, err
+	}
+	addr := urlParsed.Host
+	if urlParsed.Port() == "" {
+		addr += ":443"
+	}
+	conn, err := tls.Dial("tcp", addr, &tls.Config{InsecureSkipVerify: insecure})
+	if err != nil {
+		return nil, err
+	}
+	defer conn.Close()
+	return conn.ConnectionState().PeerCertificates, nil
+}
 
 func (cmd *applyCommand) applyDEPProfile(args []string) error {
 	flagset := flag.NewFlagSet("dep-profiles", flag.ExitOnError)
 	var (
 		flProfilePath = flagset.String("f", "", "filename of DEP profile to apply")
 		flTemplate    = flagset.Bool("template", false, "print a JSON example of a DEP profile")
+		flAnchorFile  = flagset.String("anchor", "", "filename of PEM certificate to add to trusted anchors")
+		flUseServer   = flagset.Bool("use-server-cert", false, "use the certificate presented by the server")
 	)
 	flagset.Usage = usageFor(flagset, "mdmctl apply dep-profiles [flags]")
 	if err := flagset.Parse(args); err != nil {
@@ -23,8 +46,22 @@ func (cmd *applyCommand) applyDEPProfile(args []string) error {
 	}
 
 	if *flTemplate {
-		printDEPProfileTemplate()
-		return nil
+		var anchorCerts []*x509.Certificate
+		if *flAnchorFile != "" {
+			certs, err := crypto.ReadPEMCertificatesFile(*flAnchorFile)
+			if err != nil {
+				return err
+			}
+			anchorCerts = append(anchorCerts, certs...)
+		}
+		if *flUseServer {
+			certs, err := certificatesFromURL(cmd.config.ServerURL, cmd.config.SkipVerify)
+			if err != nil {
+				return err
+			}
+			anchorCerts = append(anchorCerts, certs...)
+		}
+		return printDEPProfileTemplate(anchorCerts)
 	}
 
 	if *flProfilePath == "" {
@@ -54,9 +91,24 @@ func (cmd *applyCommand) applyDEPProfile(args []string) error {
 	return nil
 }
 
-func printDEPProfileTemplate() {
+func printDEPProfileTemplate(anchorCerts []*x509.Certificate) error {
+	var anchorCertStr string = "[]"
 
-	resp := `
+	// convert certificates into base64 encoded strings
+	// json.Marshal does this for us for byte[] arrays
+	if len(anchorCerts) > 0 {
+		var certs [][]byte
+		for _, cert := range anchorCerts {
+			certs = append(certs, cert.Raw)
+		}
+		jsonBytes, err := json.Marshal(certs)
+		if err != nil {
+			return nil
+		}
+		anchorCertStr = string(jsonBytes)
+	}
+
+	resp := fmt.Sprintf(`
 {
   "profile_name": "(Required) Human readable name",
   "url": "https://mymdm.example.org/mdm/enroll",
@@ -69,12 +121,13 @@ func printDEPProfileTemplate() {
   "support_phone_number": "(Optional) +1 408 555 1010",
   "support_email_address": "(Optional) support@example.com",
   "org_magic": "(Optional)",
-  "anchor_certs": [],
+  "anchor_certs": %s,
   "supervising_host_certs": [],
   "skip_setup_items": ["AppleID", "Android"],
   "department": "(Optional) support@example.com",
   "devices": ["SERIAL1","SERIAL2"]
 }
-`
+`, anchorCertStr)
 	fmt.Println(resp)
+	return nil
 }